refactor: add ProposeRoleForm thats used by ProposeRoleModal component
