+++ src/apollo/proposals/fragments/ProposalForm.fragment.graphql
@@ -11,6 +11,10 @@ fragment ProposalForm on Proposal {
     groupCoverPhoto {
       ...AttachedImage
     }
+
+    role {
+      ...ProposalActionRole
+    }
   }
 
   images {
